Experimental teleop? and Start changes
In the teleop our team is using I added a small change to start where the slide is extended on start to tense the string and so our pivot doesn't break again.

The experimental teleop uses 2 extra threads to handle the power of motors while the others calculate. Another thread handles telemetry as having 250 lines of code in one loop has some significant lag in value updates.